+++ yash-cli/src/startup.rs
@@ -25,6 +25,7 @@ use yash_env::io::Fd;
 use yash_env::option::Option::{Interactive, Monitor, Stdin};
 use yash_env::option::State::On;
 use yash_env::parser::IsKeyword;
+use yash_env::parser::IsName;
 use yash_env::prompt::GetPrompt;
 use yash_env::semantics::command::RunFunction;
 use yash_env::trap::RunSignalTrapIfCaught;
@@ -114,6 +115,10 @@ pub fn configure_environment(env: &mut Env, run: Run) -> Work {
 fn inject_dependencies(env: &mut Env) {
     env.any.insert(Box::new(IsKeyword(|_env, word| {
         yash_syntax::parser::lex::Keyword::from_str(word).is_ok()
+    })));
+
+    env.any.insert(Box::new(IsName(|_env, name| {
+        yash_syntax::parser::lex::is_name(name)
     })));
 
     env.any.insert(Box::new(RunReadEvalLoop(|env, config| {