+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,12 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.5] - Unreleased
+
+### Changed
+
+- Internal: Updated dependency injection to include `IsName` for variable name validation.
+
 ## [3.0.4] - 2025-11-07
 
 ### Changed